[project @ Fixed StoreTest to not fail on Mac OS X 10.5]
Mac OS X 10.5 generates a value for $TMPDIR that looks something like
this:

    /var/folders/uU/uXMNhbRIGkOhaxcg60xSGU+++TI/-Tmp-/

The '+++' is causing the DSN in Tests_Auth_OpenID_StoreTest::test_sqlitestore
to become invalid. This invalid DSN causes the SQLite driver to attempt
creating a file at a nonexistant path, resulting in a failed test. This is
fixed by wrapping the value of $TMPDIR received by the system in a call to
urlencode().