[project @ getAliasedArg() returns OpenID namespace when $aliased_key is 'ns']
This fixes an rather cryptic error when using stateless mode via the DumbStore.  The 'ns'
key can not be found in the alias/namespace mapping (its stored as the "Null Namespace"),
it must be returned explicitly. The inability to find the key in the mapping results in a
"Server Denied check_authentication" error, but the error is caused before any callback
to the server is made.

This also brings the PHP lib more in line with the ruby and python libs.