Added the following patch from the dev@openidenabled.com mailing list: http://lists.openidenabled.com/pipermail/dev/attachments/20080821/1eded806/attachment-0001.bin
Original Message:
will at willnorris.com
Thu Aug 21 18:43:28 PDT 2008
darcs patch: seperate provider and consumer xrds types

	"Currently, the five OpenID Provider service types (OpenID 1.0 - 1.2, and the two for OpenID 2.0) are combined with the one OpenID Consumer service type (return_to).  This causes a potential false positive when performing discovery on an endpoint.  If a php-openid consumer performs discovery on another consumer which has a published return_to service type, php-openid will think it is a valid OpenID Server service type and send an authentication request to the URI.  Similarly, when checking the return_to URL of a consumer, if the consumer has any of the OpenID Server service types published, a php-openid server object will try to treat one of those URIs as a published return_to URL.

	This scenario is actually fairly common... many personal blogs serve as both an OpenID Server (for the blog owner) as well as an OpenID consumer (for people who comment on the blog)."

This patch was in the form of a Darcs patch, not a normal patch.  So solve this, I applied it to the Darcs repository found on openidenabled, then created a new diff file between the original Darcs repo and the new one (with the patch applied) so that I could apply it to this git repo.

It should be noted that all 3 hunks for Discover.php and the 1st hunk for TrustRoot.php were applied successfully, but the second hunk for TrustRoot failed.  It seems that all it was doing was removing trailing endlines from the file, so I think it should be ok.