Added the following patch from the dev@openidenabled.com mailing list: http://lists.openidenabled.com/pipermail/dev/attachments/20090101/217f2590/attachment.bin
Original Message:
dev-list-openidenabled at thequod.de dev-list-openidenabled at thequod.de
Thu Jan 1 18:14:12 PST 2009
darcs patch: php-openid-urldecode-urlparamkey

	"This patch also decodes the key of URL params, which is
	required for param names like 'action[foo]' when the browser sends those
	urlencoded (Konqueror 4.2 does so).
	I would like to propose using rawurldecode instead of urldecode, too, but
	that causes the tests to fail and may be against the OpenID spec
	(the difference is only that "+" gets decoded with urldecode() but not
	with rawurldecode IIRC)"

This patch was in the form of a Darcs patch, not a normal patch.  So solve this, I applied it to the Darcs repository found on openidenabled, then created a new diff file between the original Darcs repo and the new one (with the patch applied) so that I could apply it to this git repo.

All hunks were applied successfully.