Added the following patch from the dev@openidenabled.com mailing list: http://lists.openidenabled.com/pipermail/dev/attachments/20090113/8ed067b3/attachment.bin
Original Message:
will at willnorris.com will at willnorris.com
Sun Jan 4 11:27:26 PST 2009
darcs patch: I'm not entirely sure which versions of PHP are having...

	"I've had a number of users report that $this->store is incorrectly evaluating to false, thereby preventing associations from being used.  I've been unable to reproduce this myself, but the change is minor enough."

This patch was in the form of a Darcs patch, not a normal patch.  So solve this, I applied it to the Darcs repository found on openidenabled, then created a new diff file between the original Darcs repo and the new one (with the patch applied) so that I could apply it to this git repo.

All hunks were applied successfully.