Added the following patch from the dev@openidenabled.com mailing list: http://lists.openidenabled.com/pipermail/dev/attachments/20090206/f0d91666/attachment-0001.bin
Original Message:
will at willnorris.com will at willnorris.com
Fri Feb 6 14:56:12 PST 2009
darcs patch: Properly handle Yadis discovery failure

	"The library does not curently handle failures during Yadis discovery properly. Specifically, if Yadis discovery fails, and discovery WAS being done with an XRDS document, then non-Yadis discovery should be attempted.

	The use-case I'm dealing with right now that exposes this bug is when an XRDS document location is advertised (via HTTP response header or <meta /> HTML tag), but the document is not able to be retrieved for whatever reason.  The library is failing in this case, instead of continuing to attempt non-Yadis discovery.

	This patch includes two changes:

	- change to Yadis/Yadis.php that properly detects if XRDS was used

	- change to OpenID/Discovery.php that prevents discovery from halting in the
		above example"

This patch was in the form of a Darcs patch, not a normal patch.  So solve this, I applied it to the Darcs repository found on openidenabled, then created a new diff file between the original Darcs repo and the new one (with the patch applied) so that I could apply it to this git repo.

All hunks were applied successfully.