Remove "version" from composer.json
Never add "version" to composer.json if you have a repository and can add tags. Otherwise every commit will be described as "version 3.0.0" after the first commit, and this will confuse Composer, because it is not true. Only use "version" in a context where no other ways of designating a version are available.

Tagging the version in this repo is enough for Composer to function.